% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/configurator-package.r
\docType{package}
\name{configurator-class}
\alias{configurator-class}
\alias{configurator-class-package}
\title{configurator-class}
\description{
A small set of tools for managing configuration files, formatted like INI files in Windows.  See \url{http://effbot.org/librarybook/configparser.htm} and \url{http://docs.python.org/release/2.6.8/library/configparser.html} for inspiration.
}
\examples{
\dontrun{
  Cfg <- Configurator(system.file("extdata", "example.cfg", package = "configurator")
  Cfg
  n_fleas <- Cfg$get("Bar", "fleas")
}
}
\seealso{
Other CONFIGURATOR: \code{\link{ConfiguratorRefClass_get}};
  \code{\link{ConfiguratorRefClass_parse_line}};
  \code{\link{ConfiguratorRefClass_read_file}};
  \code{\link{ConfiguratorRefClass_set}};
  \code{\link{ConfiguratorRefClass_show}};
  \code{\link{ConfiguratorRefClass_write_file}};
  \code{\link{Configurator}}
}

