% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Configurator.R
\docType{class}
\name{ConfiguratorRefClass-class}
\alias{ConfiguratorRefClass}
\alias{ConfiguratorRefClass-class}
\title{A Reference Class for configurations}
\description{
A Reference Class for configurations
}
\section{Fields}{

\describe{
\item{\code{filename}}{the name of the configuration file}

\item{\code{comment}}{the comment character}

\item{\code{container}}{a named list of the configuration sections}
}}
\section{Methods}{

\describe{
\item{\code{get(section, name, default = NULL)}}{Retrieve the specified item in the sectioned named or NULL of not found}

\item{\code{parse_line(s)}}{Given one or more raw text lines, parse the key=value pairs into the 
expected named character vector}

\item{\code{read_file(filename = .self$filename)}}{read the specified configuration file}

\item{\code{set(section, name, value)}}{Set the value of the specified item in the sectioned named}

\item{\code{write_file(filename = .self$filename, append = FALSE)}}{write the specified configuration file}
}}

